Create an Intcode class
Day 5 builds on the Intcode computer from Day 2.  To get /some/ reuse
and prepare for possible additional problems, refactor Day 2 solution to
move the Intcode execution and any operations on the Intcode computer
into a separate class.

Due to annoyances with Python importing, day solutions that require the
Intcode computer will be located in the intcode folder.  Hopefully fix
this in the future.